Closing the loops
Filled in most of the life support loops up to the present.  Given the right mix, supply runs now only need food and hydrogen.
Also added extenders for ECLSS and chemical plant.  Due to the way they work, adding a second unit with the same process quadrupled the output since each instance added both a process and a modifier.  Extenders add pseudoresources but no process so that you can have as much of a process as you need but only one button.